requirements.txt
uWSGI>=2.0.13.1,<2.1
Flask>=0.11.1,<0.12
redis>=2.10.5,<2.11
hiredis>=0.2.0,<0.3
rq>=0.6.0,<0.7
requests==2.11.0
enum34>=1.0
python-dateutil>=2.5.3,<2.6
